+++ care/security/apps.py
@@ -2,7 +2,7 @@ from django.apps import AppConfig
 from django.utils.translation import gettext_lazy as _
 
 
-class FacilityConfig(AppConfig):
+class SecurityConfig(AppConfig):
     name = "care.security"
     verbose_name = _("Security Management")
